fix: protect root folder is from being moved and split folder tests
